Fix events sessions stats (#403)
* fix the sessions count

* fix selection widget

* fix

* fix long range events

* fix stats

+++ Backend/Models/Events/EventSummary.cs
@@ -4,5 +4,5 @@ namespace Helse.Models.Events;
 
 public record EventStats(EventSummary[] Summaries, Interval[] Durations, Event[] Events);
 
-public record EventSummary(Dictionary<string, int> Data);
+public record EventSummary(Dictionary<string, double> Data);
